+++ WP90_2.c
@@ -1,5 +1,5 @@
 #include <stdio.h> //*sqrt() don't use math.h
-float sqrt(int num)
+float sqrt(float num)
 {
     float x[10000], result = 0;
     for (int i = 0; i <= 10000; i++)
@@ -12,9 +12,9 @@ float sqrt(int num)
 }
 int main()
 {
-    int num;
+    float num;
     printf("Enter number : ");
-    scanf("%d", &num);
+    scanf("%f", &num);
     printf("Result = %f", sqrt(num));
     return 0;
     return 0;
